Fix alias for lateral subqueries (#10)
When 2 tables have the same alias prefix and both are used in combination with a lateral subquery join, the outer table gets a wrong alias within the subquery:

```sql
SELECT t.Field3, sq2.f0
FROM three AS t
	INNER JOIN LATERAL (
		SELECT t.Comment f0
		FROM "two $#!" AS t
		WHERE t.OneID = t2.OneID  --t2.OneID is three.OneID but gets wrong alias
	) sq2 ON (1 = 1)
LIMIT 1
```
This results in an error:
* > panic: pq: missing FROM-clause entry for table "t2"

A workaround for this, is manually setting the alias of table "three":
```
	t3 := model.Three()
	o.GetTable().Alias = `th`
```
Which results in:
```sql
SELECT th.Field3, sq2.f0
FROM three AS th
	INNER JOIN LATERAL (
		SELECT t.Comment f0
		FROM "two $#!" AS t
		WHERE t.OneID = th.OneID
	) sq2 ON (1 = 1)
LIMIT 1
```
But this solution is fragile...

The same issue exists when 2 lateral subqueries are used, and the second wants to filter on the field of the first subquery (see added test). But no workaround is available for this situation.

Co-authored-by: Robin Knaapen <robin@ultraware.nl>
Reviewed-on: https://git.ultraware.nl/Ultraware/qb/pulls/10
Reviewed-by: Pietjan Oostra <p.oostra@ultraware.nl>
Reviewed-by: Robin Knaapen <robin@ultraware.nl>
Co-authored-by: André Mussche <andre@ultraware.nl>
Co-committed-by: André Mussche <andre@ultraware.nl>